Update Audio-Concatenator.py
Updated it so now it can process as many audio clips inside a folder without crashing, parallel processing of the files to make the process faster (Adjust max_workers parameter to a higher number depending on the power of your system). Also added a Cancel button. Also, the gui shouldn't crash because it now runs in the background.